requirements.txt
discord.py
jishaku
pymongo[srv] 
dnspython